requirements.txt
PyMySQL==1.0.3
python-slugify==8.0.1
sqlparse==0.4.3
tabulate==0.9.0
text-unidecode==1.3
